backend/requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
django-redis
redis
djangorestframework-simplejwt
django-cors-headers
python-dotenv
Pillow
